Updated forder structure, added more AboutUs Page sections
