+++ my-app/src/components/Pages/2-AboutUs-Page/1-WhoWeAre/0-WhoWeAre-Main/WhoWeAre.js
@@ -1,0 +1,22 @@
+import React from "react";
+import OurStory from "../1-OurStory/OurStory";
+import OurMission from "../2-OurMission/OurMission";
+import Message from "../3-Message/Message";
+import LearnMore from "../4-LearnMore/LearnMore";
+import Gallery from "../5-Gallery/Gallery";
+import Testimonials from "../6-Testimonials/Testimonials";
+
+const WhoWeAre = () => {
+  return (
+    <section className="WhoWeAre">
+      <h3>WhoWeAre Header</h3>
+      <OurStory />
+      <OurMission />
+      <Message />
+      <LearnMore />
+      <Gallery />
+      <Testimonials />
+    </section>
+  );
+};
+export default WhoWeAre;
