--- my-app/src/components/Pages/2-AboutUs-Page/AboutUs.css
@@ -1,3 +1,0 @@
-.AboutUs h2 {
-    color: var(--main-black);
-}
