--- my-app/src/components/Pages/2-AboutUs-Page/AboutUs.js
@@ -1,10 +1,0 @@
-import React from "react";
-
-const AboutUs = () => {
-  return (
-    <section className="AboutUs">
-      <h2>hello from About page</h2>
-    </section>
-  );
-};
-export default AboutUs;
